requirements.txt
playwrigh==1.41.1
pytes==8.0.0
allure-pytest==2.13.2